+++ driver/src/main/java/oracle/nosql/driver/httpclient/HttpClient.java
@@ -111,6 +111,9 @@ public class HttpClient {
     private final SslContext sslCtx;
     private final int handshakeTimeoutMs;
 
+    /* Enable endpoint identification by default if using SSL */
+    private boolean enableEndpointIdentification = true;
+
     private final Logger logger;
 
     /*
@@ -282,6 +285,14 @@ public class HttpClient {
         return sslCtx;
     }
 
+    public boolean isEndpointIdentificationEnabled() {
+        return enableEndpointIdentification;
+    }
+
+    public void disableEndpointIdentification() {
+        this.enableEndpointIdentification = false;
+    }
+
     public int getPort() {
         return port;
     }
@@ -366,7 +377,16 @@ public class HttpClient {
      */
     public void shutdown() {
         pool.close();
-        workerGroup.shutdownGracefully().syncUninterruptibly();
+        /*
+         * 0 means no quiet period, waiting for more tasks
+         * 5000ms is total time to wait for shutdown (should never take this
+         * long
+         *
+         * See doc:
+         * https://netty.io/4.1/api/io/netty/util/concurrent/EventExecutorGroup.html#shutdownGracefully--
+         */
+        workerGroup.shutdownGracefully(0, 5000, TimeUnit.MILLISECONDS).
+            syncUninterruptibly();
     }
 
     public Channel getChannel(int timeoutMs)