+++ driver/src/main/java/oracle/nosql/driver/iam/SecurityTokenSupplier.java
@@ -173,18 +173,7 @@ class SecurityTokenSupplier {
             }
         }
         SecurityToken token = getSecurityTokenFromIAM();
-        token.validate(minTokenLifetime);
-
-        /*
-         * Allow logging of token expiration details
-         */
-        long tokenLifetime = token.getExpiryMS() - token.getCreationTime();
-        logTrace(logger, "New security token: lifetime=" + tokenLifetime +
-                 ", expiry=" + token.getExpiryMS() + ", creation=" +
-                 token.getCreationTime());
-        if (tokenLifetime < minTokenLifetime) {
-            logTrace(logger, "token:\n" + token.getSecurityToken());
-        }
+        token.validate(minTokenLifetime, logger);
 
         return token.getSecurityToken();
     }
@@ -250,12 +239,6 @@ class SecurityTokenSupplier {
     }
 
     static class SecurityToken {
-        /* fields in security token JSON used to check validity */
-        private static final String[] FIELDS = {
-            "exp", "jwk", "n", "e",
-            ResourcePrincipalClaimKeys.COMPARTMENT_ID_CLAIM_KEY,
-            ResourcePrincipalClaimKeys.TENANT_ID_CLAIM_KEY};
-
         private final SessionKeyPairSupplier keyPairSupplier;
         private final String securityToken;
         private final Map<String, String> tokenClaims;
@@ -266,7 +249,7 @@ class SecurityTokenSupplier {
             this.securityToken = token;
             this.keyPairSupplier = supplier;
             this.creationTime = System.currentTimeMillis();
-            this.tokenClaims = parseToken(token);
+            this.tokenClaims = Utils.parseToken(token);
         }
 
         String getSecurityToken() {
@@ -289,7 +272,7 @@ class SecurityTokenSupplier {
          * Validate the token.
          * Throws IAE if any validation fails.
          */
-        void validate(long minTokenLifetime) {
+        void validate(long minTokenLifetime, Logger logger) {
             if (tokenClaims == null) {
                 throw new IllegalArgumentException(
                     "No security token cached");
@@ -308,8 +291,17 @@ class SecurityTokenSupplier {
 
             /*
              * Note: expiry check removed, as some tokens may have very short
-             * expirations.
+             * expirations, but allow logging of token expiration details
              */
+            long tokenLifetime = getExpiryMS() - getCreationTime();
+            logTrace(logger, "New security token: lifetime=" + tokenLifetime +
+                     ", expiry=" + getExpiryMS() + ", creation=" +
+                     getCreationTime());
+            if (tokenLifetime < minTokenLifetime) {
+                logTrace(logger, "Security token has shorter lifetime" +
+                         "than expected minimal token lifetime, token:\n" +
+                         getSecurityToken());
+            }
 
             /*
              * Next compare the public key inside the JWT is the same
@@ -362,83 +354,6 @@ class SecurityTokenSupplier {
 
             return actualKey.equals(expectKey);
         }
-
-        /*
-         * Parse security token JSON response get fields expiration time,
-         * modulus and public exponent of JWK, only used for security token
-         * validity check, ignores other fields.
-         *
-         * Response:
-         * {
-         *  "exp" : 1234123,
-         *  "jwk" : {
-         *    "e": "xxxx",
-         *    "n": "xxxx",
-         *    ...
-         *  }
-         *  ...
-         * }
-         */
-        static Map<String, String> parseToken(String token) {
-            if (token == null) {
-                return null;
-            }
-            String[] jwt = splitJWT(token);
-            String claimJson = new String(Base64.getUrlDecoder().decode(jwt[1]),
-                    StandardCharsets.UTF_8);
-
-            try {
-                JsonParser parser = createParser(claimJson);
-                Map<String, String> results = new HashMap<>();
-                parse(token, parser, results);
-
-                String jwkString = results.get("jwk");
-                if (jwkString == null) {
-                    return results;
-                }
-                parser = createParser(jwkString);
-                parse(token, parser, results);
-
-                return results;
-            } catch (IOException ioe) {
-                throw new IllegalStateException(
-                    "Error parsing security token "+ ioe.getMessage());
-            }
-        }
-
-        private static void parse(String json,
-                                  JsonParser parser,
-                                  Map<String, String> reults)
-            throws IOException {
-
-            if (parser.getCurrentToken() == null) {
-                parser.nextToken();
-            }
-            while (parser.getCurrentToken() != null) {
-                String field = findField(json, parser, FIELDS);
-                if (field != null) {
-                    parser.nextToken();
-                    reults.put(field, parser.getText());
-                }
-            }
-        }
-
-        private static String[] splitJWT(String jwt) {
-            final int dot1 = jwt.indexOf(".");
-            final int dot2 = jwt.indexOf(".", dot1 + 1);
-            final int dot3 = jwt.indexOf(".", dot2 + 1);
-
-            if (dot1 == -1 || dot2 == -1 || dot3 != -1) {
-                throw new IllegalArgumentException(
-                    "Given string is not in the valid access token format");
-            }
-
-            final String[] parts = new String[3];
-            parts[0] = jwt.substring(0, dot1);
-            parts[1] = jwt.substring(dot1 + 1, dot2);
-            parts[2] = jwt.substring(dot2 + 1);
-            return parts;
-        }
     }
 
     /**
@@ -457,5 +372,24 @@ class SecurityTokenSupplier {
          * throw an error otherwise.
          */
         void setMinTokenLifetime(long lifeTimeMS);
+
+        /**
+         * @hidden
+         * Prepare the provider. This method can be used to configure the
+         * service URL, does http client warm-up. This method is usually
+         * called when the NoSQLHandle is created. Does nothing by default.
+         */
+        default void prepare(NoSQLHandleConfig config) {
+        }
+
+        /**
+         * @hidden
+         * Close the provider. This method can be used to cleanup the resources
+         * used by the provider, such as close the http client. This method
+         * is usually called when the NoSQLHandle is closed. Does nothing by
+         * default.
+         */
+        default void close() {
+        }
     }
 }